+++ clients/admin-ui/src/features/common/utils.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line import/prefer-default-export
+export function capitalize(text: string): string {
+  return text.replace(/^\w/, (c) => c.toUpperCase());
+}