+++ clients/admin-ui/src/features/privacy-requests/RequestTable.tsx
@@ -1,17 +1,9 @@
-import {
-  Button,
-  Flex,
-  Table,
-  Tbody,
-  Text,
-  Th,
-  Thead,
-  Tr,
-} from "@fidesui/react";
+import { Table, Tbody, Th, Thead, Tr } from "@fidesui/react";
 import debounce from "lodash.debounce";
 import React, { useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
+import PaginationFooter from "../common/PaginationFooter";
 import {
   selectPrivacyRequestFilters,
   setPage,
@@ -58,8 +50,7 @@ const useRequestTable = () => {
 const RequestTable: React.FC<RequestTableProps> = () => {
   const { requests, total, page, size, handleNextPage, handlePreviousPage } =
     useRequestTable();
-  const startingItem = (page - 1) * size + 1;
-  const endingItem = Math.min(total, page * size);
+
   return (
     <>
       <Table size="sm">
@@ -80,36 +71,13 @@ const RequestTable: React.FC<RequestTableProps> = () => {
           ))}
         </Tbody>
       </Table>
-      <Flex justifyContent="space-between" mt={6}>
-        <Text fontSize="xs" color="gray.600">
-          {total > 0 ? (
-            <>
-              Showing {Number.isNaN(startingItem) ? 0 : startingItem} to{" "}
-              {Number.isNaN(endingItem) ? 0 : endingItem} of{" "}
-              {Number.isNaN(total) ? 0 : total} results
-            </>
-          ) : (
-            "0 results"
-          )}
-        </Text>
-        <div>
-          <Button
-            disabled={page <= 1}
-            onClick={handlePreviousPage}
-            mr={2}
-            size="sm"
-          >
-            Previous
-          </Button>
-          <Button
-            disabled={page * size >= total}
-            onClick={handleNextPage}
-            size="sm"
-          >
-            Next
-          </Button>
-        </div>
-      </Flex>
+      <PaginationFooter
+        page={page}
+        size={size}
+        total={total}
+        handleNextPage={handleNextPage}
+        handlePreviousPage={handlePreviousPage}
+      />
     </>
   );
 };