+++ clients/admin-ui/src/features/user-management/NewUserForm.tsx
@@ -18,7 +18,7 @@ import NextLink from "next/link";
 import { useRouter } from "next/router";
 import React from "react";
 
-import { USER_PRIVILEGES } from "../../constants";
+import { USER_MANAGEMENT_ROUTE, USER_PRIVILEGES } from "../../constants";
 import { isErrorWithDetail, isErrorWithDetailArray } from "../common/helpers";
 import {
   useCreateUserMutation,
@@ -76,7 +76,7 @@ const useUserForm = () => {
       );
 
       if (!("error" in updateUserPermissionsResult)) {
-        router.push("/user-management");
+        router.push(USER_MANAGEMENT_ROUTE);
       }
     },
     validate: (values) => {
@@ -246,7 +246,7 @@ const UserForm: NextPage = () => {
             </Stack>
           </Stack>
 
-          <NextLink href="/user-management" passHref>
+          <NextLink href={USER_MANAGEMENT_ROUTE} passHref>
             <Button variant="outline" mr={3} size="sm">
               Cancel
             </Button>