+++ clients/admin-ui/src/pages/user-management/new.tsx
@@ -9,6 +9,7 @@ import type { NextPage } from "next";
 import Link from "next/link";
 import React from "react";
 
+import { USER_MANAGEMENT_ROUTE } from "../../constants";
 import ProtectedRoute from "../../features/auth/ProtectedRoute";
 import NavBar from "../../features/common/NavBar";
 import NewUserForm from "../../features/user-management/NewUserForm";
@@ -24,8 +25,8 @@ const CreateNewUser: NextPage = () => (
             <Box mt={2} mb={7}>
               <Breadcrumb fontWeight="medium" fontSize="sm">
                 <BreadcrumbItem>
-                  <Link href="/user-management" passHref>
-                    <BreadcrumbLink href="/user-management">
+                  <Link href={USER_MANAGEMENT_ROUTE} passHref>
+                    <BreadcrumbLink href={USER_MANAGEMENT_ROUTE}>
                       User Management
                     </BreadcrumbLink>
                   </Link>