+++ create_test_data.py
@@ -10,6 +10,11 @@ from fidesops.core.config import config
 from fidesops.db.database import init_db
 from fidesops.db.session import get_db_session
 from fidesops.models.client import ClientDetail
+from fidesops.models.connectionconfig import (
+    AccessLevel,
+    ConnectionConfig,
+    ConnectionType,
+)
 from fidesops.models.fidesops_user import FidesopsUser
 from fidesops.models.policy import ActionType, Policy, Rule, RuleTarget
 from fidesops.models.privacy_request import PrivacyRequest, PrivacyRequestStatus
@@ -30,15 +35,17 @@ def _create_policy(
     created, policy = Policy.get_or_create(
         db=db,
         data={
-            "name": policy_key,
             "key": policy_key,
-            "client_id": client_id,
         },
     )
 
     if not created:
         # If the Policy is already created, don't create it again
         return policy
+    else:
+        policy.client_id = client_id
+        policy.name = policy_key
+        policy.save(db=db)
 
     rand = string.ascii_lowercase[:5]
     data = {}
@@ -91,6 +98,49 @@ def _create_policy(
     return policy
 
 
+def _create_connection_configs(db: orm.Session) -> None:
+    ConnectionConfig.get_or_create(
+        db=db,
+        data={
+            "key": "ci_create_test_data_datastore_connection",
+            "name": "seed datastore connection",
+            "connection_type": ConnectionType.postgres,
+            "access": AccessLevel.read,
+        },
+    )
+    ConnectionConfig.get_or_create(
+        db=db,
+        data={
+            "key": "ci_create_test_data_datastore_connection_disabled",
+            "name": "seed datastore connection disabled",
+            "connection_type": ConnectionType.mysql,
+            "access": AccessLevel.read,
+            "disabled": True,
+            "disabled_at": datetime.utcnow(),
+        },
+    )
+    ConnectionConfig.get_or_create(
+        db=db,
+        data={
+            "key": "ci_create_test_data_saas_connection",
+            "name": "seed saas connection",
+            "connection_type": ConnectionType.saas,
+            "access": AccessLevel.write,
+        },
+    )
+    ConnectionConfig.get_or_create(
+        db=db,
+        data={
+            "key": "ci_create_test_data_saas_connection_disabled",
+            "name": "seed saas connection disabled",
+            "connection_type": ConnectionType.saas,
+            "access": AccessLevel.write,
+            "disabled": True,
+            "disabled_at": datetime.utcnow(),
+        },
+    )
+
+
 def create_test_data(db: orm.Session) -> FidesopsUser:
     """Script to create test data for the Admin UI"""
     print("Seeding database with privacy requests")
@@ -137,6 +187,8 @@ def create_test_data(db: orm.Session) -> FidesopsUser:
                 },
             )
 
+    print("Adding connection configs")
+    _create_connection_configs(db)
     print("Data seeding complete!")